+++ paper/src/main/kotlin/dev/nyon/magnetic/Listeners.kt
@@ -10,13 +10,19 @@ import dev.nyon.magnetic.extensions.listen
 import io.papermc.paper.event.block.PlayerShearBlockEvent
 import org.apache.commons.lang3.mutable.MutableInt
 import org.bukkit.Bukkit
+import org.bukkit.Material
 import org.bukkit.Statistic
+import org.bukkit.block.Block
+import org.bukkit.block.BlockFace
+import org.bukkit.block.BlockState
+import org.bukkit.entity.Player
 import org.bukkit.event.Event
 import org.bukkit.event.block.BlockBreakEvent
 import org.bukkit.event.block.BlockDropItemEvent
 import org.bukkit.event.entity.EntityDeathEvent
 import org.bukkit.event.player.PlayerFishEvent
 import org.bukkit.event.player.PlayerShearEntityEvent
+import org.bukkit.inventory.ItemStack
 
 object Listeners {
 
@@ -28,7 +34,9 @@ object Listeners {
             items.removeIf { item ->
                 val copiedStack = item.clone()
                 if (player.inventory.addItem(item).isNotEmpty()) return@removeIf false
-                player.incrementStatistic(Statistic.PICKUP, copiedStack.type, copiedStack.amount)
+                if (copiedStack.amount != 0) player.incrementStatistic(
+                    Statistic.PICKUP, copiedStack.type, copiedStack.amount
+                )
                 true
             }
         }
@@ -38,9 +46,31 @@ object Listeners {
         }
     }
 
+    private val breakChainedBlocks = listOf(
+        Material.BAMBOO,
+        Material.CACTUS,
+        Material.CAVE_VINES,
+        Material.CAVE_VINES_PLANT,
+        Material.CHORUS_FLOWER,
+        Material.CHORUS_PLANT,
+        Material.KELP,
+        Material.KELP_PLANT,
+        Material.SUGAR_CANE,
+        Material.SCAFFOLDING,
+        Material.TWISTING_VINES,
+        Material.TWISTING_VINES_PLANT,
+        Material.WEEPING_VINES,
+        Material.WEEPING_VINES_PLANT
+    )
+    private val neededBlockFaces = BlockFace.entries.filter(BlockFace::isCartesian)
     fun listenForBukkitEvents() {
         listen<BlockDropItemEvent> {
             val itemStacks = items.map { it.itemStack }.toMutableList()
+
+            // Find and break all surrounding break-chained blocks that are not handled by the event
+            if (breakChainedBlocks.contains(blockState.type))
+                handleBreakChainedBlocks(block, blockState, player, itemStacks)
+
             DropEvent(itemStacks, MutableInt(), player).also(Event::callEvent)
 
             // Delete items that have been added to the inventory
@@ -110,6 +140,32 @@ object Listeners {
                     itemStacks.none { stack -> stack.isSimilar(item) }
                 }
             }
+        }
+    }
+
+    private fun handleBreakChainedBlocks(
+        block: Block,
+        blockState: BlockState,
+        player: Player,
+        itemStacks: MutableList<ItemStack>
+    ) {
+        val affectedBlocks: MutableSet<Block> = mutableSetOf()
+        fun scanSurroundingBlocks(block: Block) {
+            neededBlockFaces.forEach { face ->
+                val otherBlock = block.getRelative(face)
+                if (otherBlock == block) return@forEach
+                if (otherBlock.state.type != blockState.type) return@forEach
+                if (affectedBlocks.add(otherBlock)) scanSurroundingBlocks(otherBlock)
+            }
+        }
+
+        // Recursively scan for all matching blocks that are directly connected to the broken block
+        scanSurroundingBlocks(block)
+
+        // Add all the scanned blocks to the list of unhandled blocks and break them
+        affectedBlocks.forEach { affectedBlock ->
+            itemStacks.addAll(affectedBlock.getDrops(player.inventory.itemInMainHand, player))
+            affectedBlock.type = Material.AIR
         }
     }
 }